+++ sdk/identity/azure-identity/azure/identity/aio/_credentials/environment.py
@@ -75,8 +75,10 @@ class EnvironmentCredential(AsyncContextManager):
             expected_variables = set(EnvironmentVariables.CERT_VARS + EnvironmentVariables.CLIENT_SECRET_VARS)
             set_variables = [v for v in expected_variables if v in os.environ]
             if set_variables:
-                _LOGGER.warning(
-                    "Incomplete environment configuration. These variables are set: %s", ", ".join(set_variables)
+                _LOGGER.log(
+                    logging.INFO if set_variables == [EnvironmentVariables.AZURE_CLIENT_ID] else logging.WARNING,
+                    "Incomplete environment configuration. These variables are set: %s",
+                    ", ".join(set_variables),
                 )
             else:
                 _LOGGER.info("No environment configuration found.")