+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Other Changes
 
 - Update typing of async credentials to match the `AsyncTokenCredential` protocol.
+- If within `DefaultAzureCredential`, `EnvironmentCredential` will now use log level INFO instead of WARNING to inform users of an incomplete environment configuration.  ([#31814](https://github.com/Azure/azure-sdk-for-python/pull/31814))
 
 ## 1.14.0 (2023-08-08)